requirements.txt
python-dotenv==1.0.0
langchain==0.1.3
pinecone-client==3.0.0
openai==1.9.0
tiktoken==0.5.2
fastapi[all]
langchain-community==0.0.15
langchain-openai==0.0.3
